refactor: BlackPlayer와 WhitePlayer 공통함수 적용
