+++ src/main/kotlin/omock/model/player/Player.kt
@@ -2,8 +2,8 @@ package omock.model.player
 
 import omock.model.position.Column
 import omock.model.position.Row
+import omock.model.rule.OMockRule
 import omock.model.search.Direction
-import omock.model.search.DirectionResult
 import omock.model.search.VisitedDirectionResult
 import omock.model.stone.Stone
 
@@ -24,5 +24,16 @@ sealed class Player {
         return Stone.from(row = row, column = column)
     }
 
-    abstract fun judgementResult(visitedDirectionResult: VisitedDirectionResult): Boolean
+    fun judgementResult(visitedDirectionResult: VisitedDirectionResult): Boolean {
+        visitedDirectionResult.visited.entries.forEach { (key, result) ->
+            val reverseResultCount: Int =
+                visitedDirectionResult.visited[Direction.reverse(key)]?.count ?: OMockRule.MIN_REVERSE_COUNT
+            if (isOverMinOMockCount(reverseResultCount + result.count)) return true
+        }
+        return false
+    }
+
+    private fun isOverMinOMockCount(count: Int): Boolean {
+        return count >= OMockRule.MIN_O_MOCK_COUNT
+    }
 }
